+++ clevertap-core/src/main/java/com/clevertap/android/sdk/CleverTapAPI.java
@@ -40,6 +40,9 @@ import com.clevertap.android.sdk.featureFlags.CTFeatureFlagsController;
 import com.clevertap.android.sdk.inapp.CTLocalInApp;
 import com.clevertap.android.sdk.inapp.callbacks.FetchInAppsCallback;
 import com.clevertap.android.sdk.inapp.customtemplates.CustomTemplateContext;
+import com.clevertap.android.sdk.inapp.customtemplates.FunctionPresenter;
+import com.clevertap.android.sdk.inapp.customtemplates.JsonTemplatesProducer;
+import com.clevertap.android.sdk.inapp.customtemplates.TemplatePresenter;
 import com.clevertap.android.sdk.inapp.customtemplates.TemplateProducer;
 import com.clevertap.android.sdk.inapp.customtemplates.TemplatesManager;
 import com.clevertap.android.sdk.inapp.data.CtCacheType;
@@ -56,7 +59,6 @@ import com.clevertap.android.sdk.interfaces.NotificationHandler;
 import com.clevertap.android.sdk.interfaces.NotificationRenderedListener;
 import com.clevertap.android.sdk.interfaces.OnInitCleverTapIDListener;
 import com.clevertap.android.sdk.interfaces.SCDomainListener;
-import com.clevertap.android.sdk.network.BaseNetworkManager;
 import com.clevertap.android.sdk.network.NetworkManager;
 import com.clevertap.android.sdk.product_config.CTProductConfigController;
 import com.clevertap.android.sdk.product_config.CTProductConfigListener;
@@ -205,7 +207,12 @@ public class CleverTapAPI implements CTInboxActivity.InboxActivityListener {
      * @param spikyProxyDomain  CleverTap Spiky Proxy Domain
      * @noinspection unused
      */
-    public static void changeCredentials(String accountID, String token, String proxyDomain, String spikyProxyDomain) {
+    public static void changeCredentials(
+            String accountID,
+            String token,
+            String proxyDomain,
+            String spikyProxyDomain
+    ) {
         if (defaultConfig != null) {
             Logger.i("CleverTap SDK already initialized with accountID:" + defaultConfig.getAccountId()
                     + ", token:" + defaultConfig.getAccountToken() + ", proxyDomain: " + defaultConfig.getProxyDomain() +
@@ -216,6 +223,36 @@ public class CleverTapAPI implements CTInboxActivity.InboxActivityListener {
         }
 
         ManifestInfo.changeCredentials(accountID, token, proxyDomain, spikyProxyDomain);
+    }
+
+    /**
+     * This method is used to change the credentials of CleverTap account Id, token, proxyDomain, spikyProxyDomain programmatically
+     *
+     * @param accountID         CleverTap Account Id
+     * @param token             CleverTap Account Token
+     * @param proxyDomain       CleverTap Proxy Domain
+     * @param spikyProxyDomain  CleverTap Spiky Proxy Domain
+     * @param customHandshakeDomain  Custom handshake Domain
+     * @noinspection unused
+     */
+    public static void changeCredentials(
+            String accountID,
+            String token,
+            String proxyDomain,
+            String spikyProxyDomain,
+            String customHandshakeDomain
+    ) {
+        if (defaultConfig != null) {
+            Logger.i("CleverTap SDK already initialized with accountID:" + defaultConfig.getAccountId()
+                    + ", token:" + defaultConfig.getAccountToken() + ", proxyDomain: " + defaultConfig.getProxyDomain() +
+                    ", spikyDomain: " + defaultConfig.getSpikyProxyDomain() + ", handshakeDomain: " + defaultConfig.getCustomHandshakeDomain()
+                    + ". Cannot change credentials to accountID: " + accountID +
+                    ", token: " + token + ", proxyDomain: " + proxyDomain + ", spikyProxyDomain: " + spikyProxyDomain
+                    + "and customHandshakeDomain: " + customHandshakeDomain);
+            return;
+        }
+
+        ManifestInfo.changeCredentials(accountID, token, proxyDomain, spikyProxyDomain, customHandshakeDomain);
     }
 
     /**
@@ -1027,7 +1064,7 @@ public class CleverTapAPI implements CTInboxActivity.InboxActivityListener {
      * {@link TemplateProducer}. See {@link com.clevertap.android.sdk.inapp.customtemplates.CustomTemplate.Builder
      * CustomTemplate.Builder}. Templates must be registered before the {@link CleverTapAPI} instance, that would use
      * them, is created. A common place for this initialization is in {@link Application#onCreate()}. If your
-     * application uses multiple {@link CleverTapAPI} instance, use the {@link CleverTapInstanceConfig} within the
+     * application uses multiple {@link CleverTapAPI} instances, use the {@link CleverTapInstanceConfig} within the
      * TemplateProducer to differentiate which templates should be registered to which {@link CleverTapAPI}
      * instances.This method can be called multiple times with different TemplateProducers, however all of the
      * produced templates must have unique names.
@@ -1059,6 +1096,32 @@ public class CleverTapAPI implements CTInboxActivity.InboxActivityListener {
      */
     public static synchronized void registerCustomInAppTemplates(TemplateProducer producer) {
         TemplatesManager.register(producer);
+    }
+
+    /**
+     * Register {@link com.clevertap.android.sdk.inapp.customtemplates.CustomTemplate CustomTemplates} through a
+     * json definition. Templates must be registered before the {@link CleverTapAPI} instance, that would use
+     * them, is created. A common place for this initialization is in {@link Application#onCreate()}. If your
+     * application uses multiple {@link CleverTapAPI} instances, extend {@link JsonTemplatesProducer}
+     * with definitions for each instance and register using {@link #registerCustomInAppTemplates(TemplateProducer)}.
+     * Use the {@link CleverTapInstanceConfig} in {@link JsonTemplatesProducer#defineTemplates} to control
+     * for which instance the templates should be registered.
+     * </br>
+     * This method can be called multiple times with different json definitions and presenters,
+     * however all of the templates must have unique names.
+     *
+     * @param templatesJson A string with the json definitions of templates. See
+     *                      {@link JsonTemplatesProducer} for the json format.
+     * @param templatesPresenter A presenter for all templates in the json definitions.
+     *                           Required if there is at least one template with type "template".
+     * @param functionsPresenter A presenter for all functions in the json definitions.
+     *                           Required if there is at least one template with type "function".
+     */
+    public static synchronized void registerCustomInAppTemplates
+    (@NonNull String templatesJson,
+     @Nullable TemplatePresenter templatesPresenter,
+     @Nullable FunctionPresenter functionsPresenter) {
+        TemplatesManager.register(new JsonTemplatesProducer(templatesJson, templatesPresenter, functionsPresenter));
     }
 
     /**
@@ -1104,7 +1167,7 @@ public class CleverTapAPI implements CTInboxActivity.InboxActivityListener {
         }
 
         TemplatesManager templatesManager = coreState.getTemplatesManager();
-        BaseNetworkManager networkManager = coreState.getNetworkManager();
+        NetworkManager networkManager = coreState.getNetworkManager();
 
         getCleverTapID(x -> {
             // getCleverTapID is executed on the main thread
@@ -1365,10 +1428,10 @@ public class CleverTapAPI implements CTInboxActivity.InboxActivityListener {
     public void setSCDomainListener(SCDomainListener scDomainListener) {
         coreState.getCallbackManager().setSCDomainListener(scDomainListener);
 
-        if(coreState.getNetworkManager() != null) {
-            NetworkManager networkManager = (NetworkManager) coreState.getNetworkManager();
+        if (coreState.getNetworkManager() != null) {
+            NetworkManager networkManager = coreState.getNetworkManager();
             String domain = networkManager.getDomain(EventGroup.REGULAR);
-            if(domain != null) {
+            if (domain != null) {
                 scDomainListener.onSCDomainAvailable(getSCDomain(domain));
             }
         }
@@ -3039,6 +3102,7 @@ public class CleverTapAPI implements CTInboxActivity.InboxActivityListener {
         String accountRegion = manifest.getAccountRegion();
         String proxyDomain = manifest.getProxyDomain();
         String spikyProxyDomain = manifest.getSpikeyProxyDomain();
+        String handshakeDomain = manifest.getHandshakeDomain();
         if (accountId == null || accountToken == null) {
             Logger.i(
                     "Account ID or Account token is missing from AndroidManifest.xml, unable to create default instance");
@@ -3054,6 +3118,9 @@ public class CleverTapAPI implements CTInboxActivity.InboxActivityListener {
         }
         if (spikyProxyDomain != null && !spikyProxyDomain.trim().isEmpty()) {
             defaultInstanceConfig.setSpikyProxyDomain(spikyProxyDomain);
+        }
+        if (handshakeDomain != null && !handshakeDomain.trim().isEmpty()) {
+            defaultInstanceConfig.setCustomHandshakeDomain(handshakeDomain);
         }
         return defaultInstanceConfig;
     }