+++ clevertap-core/src/main/java/com/clevertap/android/sdk/network/http/Request.kt
@@ -2,4 +2,8 @@ package com.clevertap.android.sdk.network.http
 
 import android.net.Uri
 
-class Request(val url: Uri, val headers: Map<String, String>, val body: String?)
+class Request(
+    val url: Uri,
+    val headers: Map<String, String>,
+    val body: String?
+)