+++ clevertap-core/src/test/java/com/clevertap/android/sdk/network/api/CtApiTestProvider.kt
@@ -10,11 +10,12 @@ import org.mockito.*
 internal object CtApiTestProvider {
 
     const val DEFAULT_DOMAIN = "domain.com"
-    const val DOMAIN = "new.domain.com"
+    const val CACHED_DOMAIN = "new.domain.com"
     const val REGION = "region"
-    const val SPIKY_DOMAIN = "new-spiky.domain.com"
+    const val CACHED_SPIKY_DOMAIN = "new-spiky.domain.com"
     const val PROXY_DOMAIN = "proxy-domain.com"
     const val SPIKY_PROXY_DOMAIN = "proxy-spiky-domain.com"
+    const val CUSTOM_HANDSHAKE_DOMAIN = "custom-handshake-domain.com"
 
     const val ACCOUNT_ID = "accountId"
     const val ACCOUNT_TOKEN = "accountToken"
@@ -23,18 +24,19 @@ internal object CtApiTestProvider {
     fun provideDefaultTestCtApi(): CtApi {
 
         return CtApi(
-            MockHttpClient(),
-            DEFAULT_DOMAIN,
-            DOMAIN,
-            SPIKY_DOMAIN,
-            REGION,
-            PROXY_DOMAIN,
-            SPIKY_PROXY_DOMAIN,
-            ACCOUNT_ID,
-            ACCOUNT_TOKEN,
-            SDK_VERSION,
-            Mockito.mock(Logger::class.java),
-            "testCtApi"
+            httpClient = MockHttpClient(),
+            defaultDomain = DEFAULT_DOMAIN,
+            cachedDomain = CACHED_DOMAIN,
+            cachedSpikyDomain = CACHED_SPIKY_DOMAIN,
+            region = REGION,
+            proxyDomain = PROXY_DOMAIN,
+            spikyProxyDomain = SPIKY_PROXY_DOMAIN,
+            customHandshakeDomain = null,
+            accountId = ACCOUNT_ID,
+            accountToken = ACCOUNT_TOKEN,
+            sdkVersion = SDK_VERSION,
+            logger = Mockito.mock(Logger::class.java),
+            logTag = "testCtApi"
         )
     }
 
@@ -43,18 +45,19 @@ internal object CtApiTestProvider {
         httpClient: CtHttpClient = MockHttpClient()
     ): CtApi {
         return CtApi(
-            httpClient,
-            Constants.PRIMARY_DOMAIN,
-            null,
-            null,
-            config.accountRegion,
-            config.proxyDomain,
-            config.spikyProxyDomain,
-            config.accountId,
-            config.accountToken,
-            SDK_VERSION,
-            config.logger,
-            "testCtApi"
+            httpClient = httpClient,
+            defaultDomain = Constants.PRIMARY_DOMAIN,
+            cachedDomain = null,
+            cachedSpikyDomain = null,
+            region = config.accountRegion,
+            proxyDomain = config.proxyDomain,
+            spikyProxyDomain = config.spikyProxyDomain,
+            customHandshakeDomain = config.customHandshakeDomain,
+            accountId = config.accountId,
+            accountToken = config.accountToken,
+            sdkVersion = SDK_VERSION,
+            logger = config.logger,
+            logTag = "testCtApi"
         )
     }
 }